refactor: carregamento dinâmico das libs de geração de PDF para reduzir o tamanho do bundle
