+++ src/components/pdf-generator/pdf-generator.style.ts
@@ -1,0 +1,13 @@
+import styled from "styled-components";
+
+export const PdfGeneratorStyle = styled.div`
+  display: flex;
+  justify-content: flex-end;
+  padding: 1rem;
+  margin-top: 1rem;
+  
+  span {
+    font-size: 0.7rem;
+    font-family: "Franklin Gothic Medium", "Arial Narrow", Arial, sans-serif;
+  }
+`;
